Add last changes to authentication feature
- Add new API domain for login only (LOGIN_DOMAIN)
- Change Login POST request to use separate API (LOGIN_DOMAIN)
- Fix /logout method to redirect to home page
- Abstract DropdownMenu from LoginMenu
- Add comments to Header
