+++ src/contexts/AuthProvider.js
@@ -22,6 +22,7 @@ function AuthProvider({ children }) {
         setMatric("");
         setToken("");
         navigate("/");
+        window.location.reload();
     }
 
     return (
