+++ src/utils/Constants.js
@@ -1,2 +1,7 @@
+// Update this when using new AppScript Deployment
 export const API_DOMAIN =
-  "https://script.google.com/macros/s/AKfycbxooNFuPLMU2fSji-2U4THouVRdbWEJuT0e8Zs0JRMUh50TWveQdS4frPUEYcoeQhkDvg/exec"
+"https://script.google.com/macros/s/AKfycbxooNFuPLMU2fSji-2U4THouVRdbWEJuT0e8Zs0JRMUh50TWveQdS4frPUEYcoeQhkDvg/exec";
+
+// Do not update LOGIN_DOMAIN! Used for login POST method
+export const LOGIN_DOMAIN =
+"https://script.google.com/macros/s/AKfycbxooNFuPLMU2fSji-2U4THouVRdbWEJuT0e8Zs0JRMUh50TWveQdS4frPUEYcoeQhkDvg/exec";
